+++ src/three/math/Ellipsoid.js
@@ -1,4 +1,4 @@
-import { Vector3, Spherical, MathUtils } from 'three';
+import { Vector3, Spherical } from 'three';
 
 // Cesium / 3D tiles Spheroid:
 // - Up is Z at 90 degrees latitude
@@ -45,12 +45,6 @@ export function latitudeToSphericalPhi( latitude ) {
 const _spherical = new Spherical();
 const _norm = new Vector3();
 const _vec = new Vector3();
-
-// https://en.wikipedia.org/wiki/World_Geodetic_System
-// https://en.wikipedia.org/wiki/Flattening
-export const WGS84_RADIUS = 6378137;
-export const WGS84_FLATTENING = 1 / 298.257223563;
-export const WGS84_HEIGHT = - ( WGS84_FLATTENING * WGS84_RADIUS - WGS84_RADIUS );
 
 export class Ellipsoid {